Allow specifying API key in query string
This is less secure than other methods since it makes the API key appear
in the URL, making it easy to accidentally be leaked. But Google
Calendar supports neither the `user:password@domain` syntax, nor signing
in via OAuth, nor setting custom request headers or body, so it's the
only option to support subscribing to the calendar there.